Can no longer support 1.9.3
Ruby 1.9.3 has String#chars which return an Enumeration, but Ruby 2.0.0+ won't have this issue. We can sure run String#chars.to_a, but as these methods are intended to get called millions of times, it can slow things down, even when there's a noop. To get around this problem, I have decided not to support Ruby 1.9.3, a very very few people use that in 2021